test(aml_reader): fix incorrect fixture filename in ARIS AML reader test and corrected arisAMLreader bug in _parse_containers and _parse_nodes objects re-scaling

+++ src/pyArchimate/readers/arisAMLreader.py
@@ -324,10 +324,10 @@ def aris_reader(model: Model, root, reader=None, scale_x=0.3, scale_y=0.3, no_vi
                     color_str = "#000000"
                 if pos is not None and size is not None:
                     n = view.add(ref=None,
-                                 x=int(pos.get('Pos.X') * scale_x),
-                                 y=int(pos.get('Pos.Y') * scale_y),
-                                 w=int(size.get('Size.dX') * scale_x),
-                                 h=int(size.get('Size.dY') * scale_y),
+                                 x=int(int(pos.get('Pos.X') )* scale_x),
+                                 y=int(int(pos.get('Pos.Y')) * scale_y),
+                                 w=int(int(size.get('Size.dX')) * scale_x),
+                                 h=int(int(size.get('Size.dY')) * scale_y),
                                  node_type='Container'
                                  )
                     n.line_color = f'#{int(color_str):0>6X}'
